Revert "Set the Loading plugin to be a Core Plugin"
