+++ Tests/Clappr_Tests/PlayerTests.swift
@@ -53,13 +53,6 @@ class PlayerTests: QuickSpec {
                 player.activePlayback?.trigger(.playing)
                 expect(callbackWasCalled).to(beTrue())
             }
-
-            it("loads LoadingCorePlugin") {
-                let player = Player()
-
-                let containsPlugin = player.core?.plugins.filter { $0 is LoadingCorePlugin }.count
-                expect(containsPlugin).to(equal(1))
-            }
         }
     }
